third_party/WebKit/LayoutTests/accessibility/meter-value.html

<!DOCTYPE HTML>
<head>
<title>Test meter min max value attributes</title>
<script src="../resources/js-test.js"></script>
</head>
<body>

<meter id="meter1" value="5" min="0" max="10">5 out of 10</meter>

<script>
if (window.testRunner)
    testRunner.dumpAsText();

description("This tests that meter element attributes min, max, value are exposed to accessibility module");

if (window.accessibilityController) {
    window.axMeter1 = accessibilityController.accessibleElementById("meter1");

    shouldBe("axMeter1.minValue", "0");
    shouldBe("axMeter1.maxValue", "10");
    shouldBe("axMeter1.intValue", "5");
}
</script>
</body>
