third_party/blink/web_tests/platform/generic/accessibility/meter-value-expected.txt

This tests that meter element attributes min, max, value are exposed to accessibility module

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axMeter1.minValue is 0
PASS axMeter1.maxValue is 10
PASS axMeter1.intValue is 5
PASS successfullyParsed is true

TEST COMPLETE
